agent/code/agno2/requirements.txt
agno==1.1.13
anthropic==0.49.0
boto3==1.37.12
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
rich==13.9.4
sqlalchemy==2.0.27
pgvector==0.2.5
psycopg[binary]==3.2.6
openai>=1.0.0
ruff==0.7.4
langtrace-python-sdk==3.8.6